requirements.txt
absl-py==2.1.0
astunparse==1.6.3
bcrypt==4.2.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-MySQL==1.6.0
Flask-MySQLdb==2.0.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==25.1.24
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.70.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
keras==3.8.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
mysql-connector-python==9.2.0
mysqlclient==2.2.7
namex==0.0.8
numpy==2.0.2
opt_einsum==3.4.0
optree==0.14.0
packaging==24.2
pillow==11.1.0
protobuf==5.29.3
py3dns==4.0.2
Pygments==2.19.1
PyMySQL==1.1.1
requests==2.32.3
rich==13.9.4
setuptools==75.8.0
six==1.17.0
SQLAlchemy==2.0.37
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==2.5.0
typing_extensions==4.12.2
urllib3==2.3.0
validate-email-address==1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
WTForms==3.2.1
